requirements.txt
amqp==5.0.6
asgiref==3.3.1
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.6
django-debug-toolbar==3.2.1
django-heroku==0.0.0
django-redis==4.12.1
goto-statement==1.2
graphviz==0.16
gunicorn==20.0.4
isort==5.7.0
kiwisolver==1.3.1
kombu==5.0.2
matplotlib==3.4.0
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.1.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
